+++ src/logic/getSortedArrayFIeldIndexes.test.ts
