+++ src/logic/getSortedArrayFieldIndexes.ts
